requirements.txt
pyautogui
pillow
omniparser
openai
python-dotenv
